Introduce TableBodyCells component
